refactor: 패키지 game과 outcome 생성
