+++ src/test/java/blakcjack/domain/outcome/OutcomeTest.java
@@ -1,4 +1,4 @@
-package blakcjack.domain;
+package blakcjack.domain.outcome;
 
 import blakcjack.domain.card.Card;
 import blakcjack.domain.card.CardNumber;
